+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## [1.5.0] - 2026-02-28
 
+### New Features
+- Users can now use SAML-based authentication system to enable enterprise-grade identity federation to Opensearch domain
+
 ### General Changes
 
 - sample configurations that help users to quickly get started are in [starter_kits](starter_kits) while the examples are now on [AWS Samples](https://github.com/aws-samples/sample-config-modern-data-architecture-accelerator)
 - Users can now provide variables placeholders in the predeploy and postdeploy hook commands
 - CLI now validates `-d`, `-e`, and `-m` filter values upfront and errors if they don't match any configured domains, environments, or modules
+- CLI now supports !include tag to reference external files in YAML configurations
 - Simplified installer stack: removed CodePipeline/S3 source the need for CodeStar ARN from user, and replaced `npm install` with running `mdaa` from npmjs.org.
 
 ## [1.4.0] - 2026-01-30